split test_losses.py into individual files for each loss
