+++ tests/losses/test_divergence.py
@@ -1,0 +1,115 @@
+import math
+
+import pytest
+import torch
+
+import kornia
+
+from testing.base import BaseTester
+
+
+class TestDivergenceLoss(BaseTester):
+    @pytest.mark.parametrize(
+        "pred,target,expected",
+        [
+            (torch.full((1, 1, 2, 4), 0.125), torch.full((1, 1, 2, 4), 0.125), 0.0),
+            (torch.full((1, 7, 2, 4), 0.125), torch.full((1, 7, 2, 4), 0.125), 0.0),
+            (torch.full((1, 7, 2, 4), 0.125), torch.zeros((1, 7, 2, 4)), 0.346574),
+            (torch.zeros((1, 7, 2, 4)), torch.full((1, 7, 2, 4), 0.125), 0.346574),
+        ],
+    )
+    def test_js_div_loss_2d(self, device, dtype, pred, target, expected):
+        actual = kornia.losses.js_div_loss_2d(pred.to(device, dtype), target.to(device, dtype))
+        expected = torch.tensor(expected).to(device, dtype)
+        self.assert_close(actual, expected)
+
+    @pytest.mark.parametrize(
+        "pred,target,expected",
+        [
+            (torch.full((1, 1, 2, 4), 0.125), torch.full((1, 1, 2, 4), 0.125), 0.0),
+            (torch.full((1, 7, 2, 4), 0.125), torch.full((1, 7, 2, 4), 0.125), 0.0),
+            (torch.full((1, 7, 2, 4), 0.125), torch.zeros((1, 7, 2, 4)), 0.0),
+            (torch.zeros((1, 7, 2, 4)), torch.full((1, 7, 2, 4), 0.125), math.inf),
+        ],
+    )
+    def test_kl_div_loss_2d(self, device, dtype, pred, target, expected):
+        actual = kornia.losses.kl_div_loss_2d(pred.to(device, dtype), target.to(device, dtype))
+        expected = torch.tensor(expected).to(device, dtype)
+        self.assert_close(actual, expected)
+
+    @pytest.mark.parametrize(
+        "pred,target,expected",
+        [
+            (torch.full((1, 1, 2, 4), 0.125), torch.full((1, 1, 2, 4), 0.125), torch.full((1, 1), 0.0)),
+            (torch.full((1, 7, 2, 4), 0.125), torch.full((1, 7, 2, 4), 0.125), torch.full((1, 7), 0.0)),
+            (torch.full((1, 7, 2, 4), 0.125), torch.zeros((1, 7, 2, 4)), torch.full((1, 7), 0.0)),
+            (torch.zeros((1, 7, 2, 4)), torch.full((1, 7, 2, 4), 0.125), torch.full((1, 7), math.inf)),
+        ],
+    )
+    def test_kl_div_loss_2d_without_reduction(self, device, dtype, pred, target, expected):
+        actual = kornia.losses.kl_div_loss_2d(pred.to(device, dtype), target.to(device, dtype), reduction="none")
+        self.assert_close(actual, expected.to(device, dtype))
+
+    @pytest.mark.parametrize(
+        "pred,target,expected",
+        [
+            (torch.full((1, 1, 2, 4), 0.125), torch.full((1, 1, 2, 4), 0.125), 0.0),
+            (torch.full((1, 7, 2, 4), 0.125), torch.full((1, 7, 2, 4), 0.125), 0.0),
+            (torch.full((1, 7, 2, 4), 0.125), torch.zeros((1, 7, 2, 4)), 0.0),
+            (torch.zeros((1, 7, 2, 4)), torch.full((1, 7, 2, 4), 0.125), math.inf),
+        ],
+    )
+    def test_noncontiguous_kl(self, device, dtype, pred, target, expected):
+        pred = pred.to(device, dtype).view(pred.shape[::-1]).transpose(-2, -1)
+        target = target.to(device, dtype).view(target.shape[::-1]).transpose(-2, -1)
+        actual = kornia.losses.kl_div_loss_2d(pred, target)
+        expected = torch.tensor(expected).to(device, dtype)
+        self.assert_close(actual, expected)
+
+    @pytest.mark.parametrize(
+        "pred,target,expected",
+        [
+            (torch.full((1, 1, 2, 4), 0.125), torch.full((1, 1, 2, 4), 0.125), 0.0),
+            (torch.full((1, 7, 2, 4), 0.125), torch.full((1, 7, 2, 4), 0.125), 0.0),
+            (torch.full((1, 7, 2, 4), 0.125), torch.zeros((1, 7, 2, 4)), 0.303251),
+            (torch.zeros((1, 7, 2, 4)), torch.full((1, 7, 2, 4), 0.125), 0.303251),
+        ],
+    )
+    def test_noncontiguous_js(self, device, dtype, pred, target, expected):
+        pred = pred.to(device, dtype).view(pred.shape[::-1]).transpose(-2, -1)
+        target = target.to(device, dtype).view(target.shape[::-1]).transpose(-2, -1)
+        actual = kornia.losses.js_div_loss_2d(pred, target)
+        expected = torch.tensor(expected).to(device, dtype)
+        self.assert_close(actual, expected)
+
+    def test_gradcheck_kl(self, device):
+        dtype = torch.float64
+        pred = torch.rand(1, 1, 10, 16, device=device, dtype=dtype)
+        target = torch.rand(1, 1, 10, 16, device=device, dtype=dtype)
+
+        # evaluate function gradient
+        self.gradcheck(kornia.losses.kl_div_loss_2d, (pred, target))
+
+    def test_gradcheck_js(self, device):
+        dtype = torch.float64
+        pred = torch.rand(1, 1, 10, 16, device=device, dtype=dtype)
+        target = torch.rand(1, 1, 10, 16, device=device, dtype=dtype)
+
+        # evaluate function gradient
+        self.gradcheck(kornia.losses.js_div_loss_2d, (pred, target))
+
+    def test_dynamo_kl(self, device, dtype, torch_optimizer):
+        pred = torch.full((1, 1, 2, 4), 0.125, dtype=dtype, device=device)
+        target = torch.full((1, 1, 2, 4), 0.125, dtype=dtype, device=device)
+        args = (pred, target)
+        op = kornia.losses.kl_div_loss_2d
+        op_optimized = torch_optimizer(op)
+        self.assert_close(op(*args), op_optimized(*args), rtol=0, atol=1e-5)
+
+    def test_dynamo_js(self, device, dtype, torch_optimizer):
+        pred = torch.full((1, 1, 2, 4), 0.125, dtype=dtype, device=device)
+        target = torch.full((1, 1, 2, 4), 0.125, dtype=dtype, device=device)
+        args = (pred, target)
+        op = kornia.losses.js_div_loss_2d
+        op_optimized = torch_optimizer(op)
+        self.assert_close(op(*args), op_optimized(*args), rtol=0, atol=1e-5)
